examples/nosetest_example/requirements.txt
nose
pagium
pyhamcrest
nose-parameterized
nose-allure-plugin
